+++ src/Headless.Payments.Paymob.Services/CashIn/Requests/PaymobCardCashInRequest.cs
@@ -19,7 +19,7 @@ namespace Headless.Payments.Paymob.Services.CashIn.Requests;
 public sealed record PaymobCardCashInRequest(
     decimal Amount,
     PaymobCashInCustomerData Customer,
-    int CardIntegrationId,
+    long CardIntegrationId,
     string? MerchantOrderId = null,
     int ExpirationSeconds = 60 * 60,
     string IframeSrc = "75432"
